+++ src/commands/Settings Commands/bot2bot.js
@@ -4,9 +4,9 @@
 
 // codebeat:disable[LOC,ABC,BLOCK_NESTING,ARITY]
 /* eslint-disable no-undef */
-const colors = require("../core/colors");
-const db = require("../core/db");
-const logger = require("../core/logger");
+const colors = require("../../core/colors");
+const db = require("../../core/db");
+const logger = require("../../core/logger");
 const discord = require("discord.js");
 
 // -------------
