+++ src/commands/Settings Commands/debug.js
@@ -4,17 +4,16 @@
 
 // codebeat:disable[LOC,ABC,BLOCK_NESTING,ARITY]
 /* eslint-disable no-undef */
-const colors = require("../core/colors");
-const db = require("../core/db");
-const logger = require("../core/logger");
+const colors = require("../../core/colors");
+const db = require("../../core/db");
+const logger = require("../../core/logger");
 const discord = require("discord.js");
 
-// -------------
-// Command Code
-// -------------
+// -----------------------
+// Command code
+// -----------------------
 
 module.exports.run = function(data)
-
 {
    // -------------------------------
    // Command allowed by admins only
@@ -23,7 +22,7 @@ module.exports.run = function(data)
    if (data.message.isAdmin === false)
    {
       data.color = "warn";
-      data.text = ":cop:  This command is reserved for server adminis.";
+      data.text = ":cop:  This command is reserved for server admins.";
 
       // -------------
       // Send message
@@ -33,15 +32,15 @@ module.exports.run = function(data)
    }
 
    // --------------------------------
-   // Error if embed param is missing
+   // Error if debug param is missing
    // --------------------------------
 
    if (!data.cmd.params)
    {
       data.color = "error";
       data.text =
-         ":warning:  Missing `embed` parameter. Use `" +
-         `${data.config.translateCmdShort} help embed\` to learn more.`;
+         ":warning:  Missing `debug` parameter. Use `" +
+         `${data.config.translateCmdShort} help debug\` to learn more.`;
 
       // -------------
       // Send message
@@ -56,35 +55,65 @@ module.exports.run = function(data)
 
    if (data.message.isAdmin)
    {
-      embed(data);
+      debug(data);
    }
 };
 
+
 // -------------------------------
-// embed varible command handaler
+// debug varible command handler
 // -------------------------------
 
-const embed = function(data)
+const debug = function(data)
 {
    const commandVariable1 = data.cmd.params.split(" ")[0].toLowerCase();
 
-   if (commandVariable1 === "on" || commandVariable1 === "off")
+   if (commandVariable1 === "on")
    {
       console.log(commandVariable1);
-      return db.updateEmbedVar(
+      return db.updateWebhookVar(
          data.message.channel.guild.id,
-         commandVariable1,
+         commandVariable1, //This would be the Webhook ID
+         commandVariable1, //this would be the Webhook Token
+         true,
          function(err)
          {
             if (err)
             {
                return logger("error", err);
             }
-            var output =
-            "**```Embedded Translation```**\n" +
-            `Embedded Message translation is now turned : ${commandVariable1}\n\n`;
+            var outputgh =
+            "**```Start Debug mode```**\n" +
+            `Debug mode has been Started. \n` +
+            `Error Logs will be output to this channel \n\n`;
             data.color = "info";
-            data.text = output;
+            data.text = outputgh;
+
+            // -------------
+            // Send message
+            // -------------
+
+            return sendMessage(data);
+         }
+      );
+   }
+   else if (commandVariable1 === "off")
+   {
+      console.log(commandVariable1);
+      return db.removeWebhook(
+         data.message.channel.guild.id,
+         function(err)
+         {
+            if (err)
+            {
+               return logger("error", err);
+            }
+            var outputoc =
+          "**```Stop Debug mode```**\n" +
+          `Debug mode has been Stopped. \n` +
+          `Error logs will not be shown.\n\n`;
+            data.color = "info";
+            data.text = outputoc;
 
             // -------------
             // Send message
@@ -98,7 +127,7 @@ const embed = function(data)
    data.color = "error";
    data.text =
       ":warning:  **`" + commandVariable1 +
-      "`** is not a valid embed option.\n";
+      "`** is not a valid debug option.\n";
 
    // -------------
    // Send message
