+++ src/commands/Translation Commands/translate.auto.js
