+++ src/commands/Translation Commands/translate.react.js
