+++ src/commands/Translation Commands/translate.stop.js
