+++ src/commands/Utility Commands/donate.js
@@ -4,7 +4,7 @@
 
 // codebeat:disable[LOC,ABC,BLOCK_NESTING,ARITY]
 /* eslint-disable no-undef */
-const colors = require("../core/colors");
+const colors = require("../../core/colors");
 const discord = require("discord.js");
 
 // -------------
